+++ src/Commands/PowerPlatform/PowerAutomate/GetFlowOwner.cs
@@ -1,0 +1,39 @@
+﻿using PnP.PowerShell.Commands.Base;
+using PnP.PowerShell.Commands.Base.PipeBinds;
+using PnP.PowerShell.Commands.Model.PowerPlatform.PowerAutomate;
+using PnP.PowerShell.Commands.Utilities.REST;
+using System.Management.Automation;
+
+namespace PnP.PowerShell.Commands.PowerPlatform.PowerAutomate
+{
+    [Cmdlet(VerbsCommon.Get, "PnPFlowOwner")]
+    public class GetFlowOwner : PnPAzureManagementApiCmdlet
+    {
+        [Parameter(Mandatory = true)]
+        public PowerPlatformEnvironmentPipeBind Environment;
+
+        [Parameter(Mandatory = true)]
+        public PowerAutomateFlowPipeBind Identity;
+
+        [Parameter(Mandatory = false)]
+        public SwitchParameter AsAdmin;
+
+        protected override void ExecuteCmdlet()
+        {
+            var environmentName = Environment.GetName();
+            if (string.IsNullOrEmpty(environmentName))
+            {
+                throw new PSArgumentException("Environment not found.", nameof(Environment));
+            }
+
+            var flowName = Identity.GetName();
+            if (string.IsNullOrEmpty(flowName))
+            {
+                throw new PSArgumentException("Flow not found.", nameof(Identity));
+            }
+
+            var flowOwners = GraphHelper.GetResultCollectionAsync<FlowPermission>(Connection, $"https://management.azure.com/providers/Microsoft.ProcessSimple{(AsAdmin ? "/scopes/admin" : "")}/environments/{environmentName}/flows/{flowName}/permissions?api-version=2016-11-01", AccessToken).GetAwaiter().GetResult();
+            WriteObject(flowOwners, true);
+        }
+    }
+}